Moving common defi deploy code to utils
