+++ test/makerV2Manager_loan.js
@@ -1,14 +1,14 @@
 const ethers = require("ethers");
-const etherlime = require("etherlime-lib");
-const { bigNumToBytes32 } = require("../utils/utilities.js");
+const {
+  bigNumToBytes32, ETH_TOKEN, RAY, ETH_PER_DAI, ETH_PER_MKR,
+} = require("../utils/utilities.js");
+const { deployMaker, deployUniswap } = require("../utils/defi-deployer");
 
-const { parseEther, bigNumberify, formatBytes32String } = ethers.utils;
+const { parseEther, formatBytes32String } = ethers.utils;
 const { HashZero, AddressZero } = ethers.constants;
 
+/* global accounts */
 const TestManager = require("../utils/test-manager");
-const UniswapFactory = require("../lib/uniswap/UniswapFactory");
-const UniswapExchange = require("../lib/uniswap/UniswapExchange");
-const ScdMcdMigration = require("../build/ScdMcdMigration");
 const Registry = require("../build/ModuleRegistry");
 const MakerV1Manager = require("../build/MakerManager");
 const MakerV2Manager = require("../build/MakerV2Manager");
@@ -19,29 +19,6 @@ const GuardianStorage = require("../build/GuardianStorage");
 const TransferStorage = require("../build/TransferStorage");
 const TransferManager = require("../build/TransferManager");
 const TokenPriceProvider = require("../build/TokenPriceProvider");
-const DSValue = require("../build/DSValue");
-const DSToken = require("../build/DSToken");
-const Dai = require("../build/Dai");
-const Vox = require("../build/SaiVox");
-const Tub = require("../build/SaiTub");
-const WETH = require("../build/WETH9");
-const Vat = require("../build/Vat");
-const Pot = require("../build/Pot");
-const Jug = require("../build/Jug");
-const CdpManager = require("../build/DssCdpManager");
-const GemJoin = require("../build/GemJoin");
-const DaiJoin = require("../build/DaiJoin");
-
-const ETH_TOKEN = "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE";
-const RAY = bigNumberify("1000000000000000000000000000"); // 10**27
-const WAD = bigNumberify("1000000000000000000"); // 10**18
-const RAD = RAY.mul(WAD);
-const USD_PER_DAI = RAY; // 1 DAI = 1 USD
-const USD_PER_ETH = WAD.mul(100); // 1 ETH = 100 USD
-const USD_PER_MKR = WAD.mul(400); // 1 MKR = 400 USD
-const ETH_PER_MKR = WAD.mul(USD_PER_MKR).div(USD_PER_ETH); // 1 MKR = 4 ETH
-const ETH_PER_DAI = WAD.mul(USD_PER_DAI).div(RAY).mul(WAD).div(USD_PER_ETH); // 1 DAI = 0.01 ETH
-const MAT = RAY.mul(3).div(2); // collateralization ratio = 150%
 
 describe("MakerV2 Vaults", function () {
   this.timeout(100000);
@@ -49,8 +26,8 @@ describe("MakerV2 Vaults", function () {
   const manager = new TestManager();
   const { deployer } = manager;
 
-  const infrastructure = global.accounts[0].signer;
-  const owner = global.accounts[1].signer;
+  const infrastructure = accounts[0].signer;
+  const owner = accounts[1].signer;
 
   let sai;
   let dai;
@@ -74,139 +51,22 @@ describe("MakerV2 Vaults", function () {
   let uniswapFactory;
 
   before(async () => {
-    //
-    // Deploy and setup SCD
-    //
-    sai = await deployer.deploy(DSToken, {}, formatBytes32String("SAI"));
-    dai = await deployer.deploy(Dai, {}, 42);
-    gov = await deployer.deploy(DSToken, {}, formatBytes32String("MKR"));
-    weth = await deployer.deploy(WETH);
-    const vox = await deployer.deploy(Vox, {}, USD_PER_DAI);
-    const sin = await deployer.deploy(DSToken, {}, formatBytes32String("SIN"));
-    const skr = await deployer.deploy(DSToken, {}, formatBytes32String("PETH"));
-    const pip = await deployer.deploy(DSValue);
-    const pep = await deployer.deploy(DSValue);
-    const tub = await deployer.deploy(Tub, {},
-      sai.contractAddress,
-      sin.contractAddress,
-      skr.contractAddress,
-      weth.contractAddress,
-      gov.contractAddress,
-      pip.contractAddress,
-      pep.contractAddress,
-      vox.contractAddress,
-      infrastructure.address);
-    // let the Tub mint PETH and DAI
-    await skr.setOwner(tub.contractAddress);
-    await sai.setOwner(tub.contractAddress);
-    // setup USD/ETH oracle with a convertion rate of 100 USD/ETH
-    await pip.poke(`0x${USD_PER_ETH.toHexString().slice(2).padStart(64, "0")}`);
-    // setup USD/MKR oracle with a convertion rate of 400 USD/MKR
-    await pep.poke(`0x${USD_PER_MKR.toHexString().slice(2).padStart(64, "0")}`);
-    // set the total DAI debt ceiling to 50,000 DAI
-    await tub.mold(formatBytes32String("cap"), parseEther("50000"));
-    // set the collateralization ratio to 150%
-    await tub.mold(formatBytes32String("mat"), MAT);
-    // set the governance fee to 7.5% APR
-    await tub.mold(formatBytes32String("fee"), "1000000002293273137447730714");
+    // Deploy Maker
+    const mk = await deployMaker(deployer, infrastructure);
+    [sai, dai, gov, bat, weth, vat, batJoin, cdpManager, pot, jug, migration] = [
+      mk.sai, mk.dai, mk.gov, mk.bat, mk.weth, mk.vat, mk.batJoin, mk.cdpManager, mk.pot, mk.jug, mk.migration,
+    ];
+    const { wethJoin, tub } = mk;
 
+    // Deploy Uniswap
+    const uni = await deployUniswap(deployer, manager, infrastructure, [gov, dai], [ETH_PER_MKR, ETH_PER_DAI]);
+    uniswapFactory = uni.uniswapFactory;
 
-    //
-    // Deploy and setup MCD
-    //
-    vat = await deployer.deploy(Vat);
-    // Setting the debt ceiling
-    await vat["file(bytes32,uint256)"](formatBytes32String("Line"), "138000000000000000000000000000000000000000000000000000");
-    cdpManager = await deployer.deploy(CdpManager, {}, vat.contractAddress);
-    pot = await deployer.deploy(Pot, {}, vat.contractAddress);
-    await vat.rely(pot.contractAddress);
-    jug = await deployer.deploy(Jug, {}, vat.contractAddress);
-    await vat.rely(jug.contractAddress);
-
-    // SAI collateral setup
-    const saiIlk = formatBytes32String("SAI");
-    await jug.init(saiIlk);
-    await vat.init(saiIlk);
-    await vat.file(saiIlk, formatBytes32String("spot"), "100000000000000000000000000000000000000000000000000");
-    await vat.file(saiIlk, formatBytes32String("line"), "100000000000000000000000000000000000000000000000000000");
-    await vat.file(saiIlk, formatBytes32String("dust"), "0");
-    const saiJoin = await deployer.deploy(GemJoin, {}, vat.contractAddress, saiIlk, sai.contractAddress);
-    await vat.rely(saiJoin.contractAddress);
-    // WETH collateral setup
-    const wethIlk = formatBytes32String("ETH-A");
-    await jug.init(wethIlk);
-    await vat.init(wethIlk);
-    await vat.file(wethIlk, formatBytes32String("spot"), "88050000000000000000000000000");
-    await vat.file(wethIlk, formatBytes32String("line"), "50000000000000000000000000000000000000000000000000000");
-    await vat.file(wethIlk, formatBytes32String("dust"), "20000000000000000000000000000000000000000000000");
-    const wethJoin = await deployer.deploy(GemJoin, {}, vat.contractAddress, wethIlk, weth.contractAddress);
-    await vat.rely(wethJoin.contractAddress);
-    // BAT collateral setup
-    bat = await deployer.deploy(DSToken, {}, formatBytes32String("BAT"));
-    const batIlk = formatBytes32String("BAT-A");
-    await jug.init(batIlk);
-    await vat.init(batIlk);
-    await vat.file(batIlk, formatBytes32String("spot"), "88050000000000000000000000000");
-    await vat.file(batIlk, formatBytes32String("line"), "50000000000000000000000000000000000000000000000000000");
-    await vat.file(batIlk, formatBytes32String("dust"), "20000000000000000000000000000000000000000000000");
-    batJoin = await deployer.deploy(GemJoin, {}, vat.contractAddress, batIlk, bat.contractAddress);
-    await vat.rely(batJoin.contractAddress);
-
-    // DAI debt setup
-    const daiJoin = await deployer.deploy(DaiJoin, {}, vat.contractAddress, dai.contractAddress);
-    // allow daiJoin to mint DAI
-    await dai.rely(daiJoin.contractAddress);
-    // give daiJoin some internal DAI in the vat
-    await vat.suck(daiJoin.contractAddress, daiJoin.contractAddress, RAD.mul(1000000));
-
-    // Deploy and setup SCD to MCD Migration
-    migration = await deployer.deploy(
-      ScdMcdMigration,
-      {},
-      tub.contractAddress,
-      cdpManager.contractAddress,
-      saiJoin.contractAddress,
-      wethJoin.contractAddress,
-      daiJoin.contractAddress,
-    );
-    // Setting up the common migration vault used by ScdMcdMigration
-    const initialSaiAmountInMigrationVault = parseEther("1000");
-    await sai["mint(address,uint256)"](infrastructure.address, initialSaiAmountInMigrationVault);
-    await sai.from(infrastructure).approve(migration.contractAddress, initialSaiAmountInMigrationVault);
-    await migration.from(infrastructure).swapSaiToDai(initialSaiAmountInMigrationVault, { gasLimit: 2000000 });
-
-    //
-    // Deploy and setup Uniswap
-    //
-    uniswapFactory = await deployer.deploy(UniswapFactory);
-    const uniswapTemplateExchange = await deployer.deploy(UniswapExchange);
-    await uniswapFactory.initializeFactory(uniswapTemplateExchange.contractAddress);
-    const ethLiquidity = parseEther("10");
-    // MKR
-    await uniswapFactory.from(infrastructure).createExchange(gov.contractAddress);
-    const mkrExchange = await etherlime.ContractAt(UniswapExchange, await uniswapFactory.getExchange(gov.contractAddress));
-    const mkrLiquidity = ethLiquidity.mul(WAD).div(ETH_PER_MKR);
-    await gov["mint(address,uint256)"](infrastructure.address, mkrLiquidity);
-    await gov.from(infrastructure).approve(mkrExchange.contractAddress, mkrLiquidity);
-    let timestamp = await manager.getTimestamp(await manager.getCurrentBlock());
-    await mkrExchange.from(infrastructure).addLiquidity(1, mkrLiquidity, timestamp + 300, { value: ethLiquidity, gasLimit: 150000 });
-    // DAI
-    await uniswapFactory.from(infrastructure).createExchange(dai.contractAddress);
-    const daiExchange = await etherlime.ContractAt(UniswapExchange, await uniswapFactory.getExchange(dai.contractAddress));
-    const daiLiquidity = ethLiquidity.mul(WAD).div(ETH_PER_DAI);
-    await dai["mint(address,uint256)"](infrastructure.address, daiLiquidity);
-    await dai.from(infrastructure).approve(daiExchange.contractAddress, daiLiquidity);
-    timestamp = await manager.getTimestamp(await manager.getCurrentBlock());
-    await daiExchange.from(infrastructure).addLiquidity(1, daiLiquidity, timestamp + 300, { value: ethLiquidity, gasLimit: 150000 });
-
-    //
     // Deploy MakerV2Manager
-    //
     registry = await deployer.deploy(Registry);
     guardianStorage = await deployer.deploy(GuardianStorage);
     makerRegistry = await deployer.deploy(MakerRegistry);
     await makerRegistry.addCollateral(wethJoin.contractAddress);
-
     makerV2 = await deployer.deploy(
       MakerV2Manager,
       {},
@@ -220,8 +80,17 @@ describe("MakerV2 Vaults", function () {
       { gasLimit: 8000000 },
     );
 
+    // Deploy MakerManager
+    makerV1 = await deployer.deploy(
+      MakerV1Manager,
+      {},
+      registry.contractAddress,
+      guardianStorage.contractAddress,
+      tub.contractAddress,
+      uniswapFactory.contractAddress,
+    );
 
-    // ???
+    // Deploy TransferManager
     const priceProvider = await deployer.deploy(TokenPriceProvider, {}, AddressZero);
     const transferStorage = await deployer.deploy(TransferStorage);
     transferManager = await deployer.deploy(TransferManager, {},
@@ -233,15 +102,6 @@ describe("MakerV2 Vaults", function () {
       3600,
       10000,
       AddressZero);
-
-    makerV1 = await deployer.deploy(
-      MakerV1Manager,
-      {},
-      AddressZero,
-      guardianStorage.contractAddress,
-      tub.contractAddress,
-      uniswapFactory.contractAddress,
-    );
   });
 
   beforeEach(async () => {
@@ -250,7 +110,6 @@ describe("MakerV2 Vaults", function () {
     walletAddress = wallet.contractAddress;
     await infrastructure.sendTransaction({ to: walletAddress, value: parseEther("0.6") });
   });
-
 
   async function getTestAmounts(tokenAddress) {
     const tokenAddress_ = (tokenAddress === ETH_TOKEN) ? weth.contractAddress : tokenAddress;
